[CREATE] 의존성 역전
