+++ ChooseImage/Sources/ImageVerificationImp/Class/ImageManager.swift
