+++ ChooseImage/Sources/Loading/LoadingInterface.swift
@@ -1,0 +1,17 @@
+//
+//  File.swift
+//  
+//
+//  Created by Ji-hoon Ahn on 2022/04/11.
+//
+
+import UIKit
+import RIBs
+
+public protocol LoadingModalListener: AnyObject {
+    func didClearTrainingMachine(image : UIImage)
+}
+
+public protocol LoadingModalBuildable: Buildable {
+    func build(withListener listener: LoadingModalListener, originerImage: UIImage, pieceImage: UIImage) -> ViewableRouting
+}
