+++ ChooseImage/Sources/LoadingImp/LoadingModalViewController.swift
