+++ Picture.xcodeproj/xcuserdata/ji-hoonahn.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -154,7 +154,7 @@
 		<key>Picture.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>1</integer>
 		</dict>
 		<key>Rx (Playground) 1.xcscheme</key>
 		<dict>
