+++ Picture/Source/AppRoot/AppRootComponent.swift
@@ -8,9 +8,13 @@
 import RIBs
 import RIBsUtil
 import ImageVerification
+import ImageVerificationImp
 import ListHome
+import ListHomeImp
 import ChooseImage
+import ChooseImageImp
 import Loading
+import LoadingImp
 import ImageRepository
 import AppHome
